simplified and corrected signing fulfilment, composes signing methods on Operator with format agmostic aigning by TransactingPower, includes cohort membership verification.
